+++ lib/models/observability_response.dart
@@ -5,6 +5,10 @@ class ObservabilityResponse {
   Object? responseBody;
   ObservabilityResponse({this.responseBody, this.status});
 
+  static ObservabilityResponse fromApi(Map<String, dynamic> json) {
+    return ObservabilityResponse();
+  }
+
   String get printIcon =>
       status == ResponseStatus.succes ? "✅✅✅✅✅✅✅✅✅✅" : "❌❌❌❌❌❌❌❌❌❌";